Update rainstate.py
Change OFF and ON values to match HA expected binary_sensor values. 

Breaking Change: was previously "DRY" and "WET", is now "OFF" and "ON" when setting raw_data=false. When setting raw_data=true, it will be "OFF" and "ON" on the HA specific topics and "0" or "1" on the generic MQTT topic.